удалил LoginWindow
